Remove types (#20)
* remove touples and crazy types
* readme

+++ Sources/PSQLKit/Directives/From/DeleteDirective.swift
@@ -1,20 +1,23 @@
 import Foundation
 import SQLKit
 
-public struct DeleteDirective<Content>: SQLExpression where Content: FromSQLExpression {
-    let content: Content
+public struct DeleteDirective: SQLExpression {
+    let content: [FromSQLExpression]
     
-    public init(@FromBuilder builder: () -> Content) {
+    public init(@FromBuilder builder: () -> [FromSQLExpression]) {
         self.content = builder()
     }
     
     public func serialize(to serializer: inout SQLSerializer) {
-        serializer.write("DELETE FROM")
-        serializer.writeSpace()
-        content.fromSqlExpression.serialize(to: &serializer)
+        if !content.isEmpty {
+            serializer.write("DELETE FROM")
+            serializer.writeSpace()
+            SQLList(content.map(\.fromSqlExpression))
+                .serialize(to: &serializer)
+        }
     }
 }
 
 extension DeleteDirective: QuerySQLExpression {
-    public var querySqlExpression: some SQLExpression { self }
+    public var querySqlExpression: SQLExpression { self }
 }
